requirements.txt
streamlit==1.39.0
seaborn==0.13.2
python-dotenv==1.0.1